Redesign the party codex to use tags
