+++ lang/en.json
@@ -1931,6 +1931,14 @@
 		"OptionArcanumPulse": "Arcanum Pulse",
 		"OptionArcanumPulseHint": "Whether to add pulses to arcanums as per the playtest materials.",
 		"Summon": "Summon",
-		"Codex": "Codex"
+		"Codex": "Codex",
+		"Tags": "Tags",
+		"CODEX": {
+			"Character": "Character",
+			"Location": "Location",
+			"Faction": "Faction",
+			"Event": "Event",
+			"Lore": "Lore"
+		}
 	}
 }
