+++ module/helpers/handlebars.mjs
@@ -237,6 +237,7 @@ export const FUHandlebars = Object.freeze({
 		Handlebars.registerHelper('pfuBadge', badge);
 		Handlebars.registerHelper('pfuItemAnchor', itemAnchor);
 		Handlebars.registerHelper('pfuCompendium', compendium);
+		Handlebars.registerHelper('pfuArrayField', arrayField);
 	},
 });
 
@@ -463,6 +464,35 @@ function compendium(tab, options) {
 			? template({
 					tab: tab,
 					...options,
+				})
+			: '';
+	return new Handlebars.SafeString(html);
+}
+
+/**
+ * @typedef FUArrayFieldOptions
+ * @property {*[]} array
+ * @property {String} label
+ * @property {String} path
+ * @property {'string'} type
+ * @property {String[]|undefined} options
+ */
+
+/**
+ * @param {FUArrayFieldOptions} options
+ * @returns {Handlebars.SafeString}
+ */
+function arrayField(options) {
+	options = options.hash;
+	const template = Handlebars.partials[systemTemplatePath('common/array-field')];
+	const html =
+		typeof template === 'function'
+			? template({
+					array: options.array,
+					label: options.label,
+					path: options.path,
+					type: options.type,
+					options: options.options,
 				})
 			: '';
 	return new Handlebars.SafeString(html);
