+++ styles/css/actor/party.css
@@ -288,37 +288,6 @@
 		max-width: 100%;
 		overflow-x: hidden;
 
-		.codex-nav {
-			display: flex;
-			flex-direction: row;
-			gap: 1rem;
-		}
-
-		.codex-nav__tab {
-			font-size: large;
-			background: transparent;
-			border-color: transparent;
-			box-shadow: none;
-			color: var(--pfu-color-control-content, inherit);
-			opacity: 0.6;
-			transition:
-				opacity 0.15s ease,
-				background 0.15s ease,
-				border-color 0.15s ease;
-
-			&:hover:not(.active) {
-				opacity: 0.85;
-			}
-
-			&.active {
-				opacity: 1;
-				background: radial-gradient(ellipse at 50% 0%, rgba(255, 255, 255, 0.22) 0%, rgba(255, 255, 255, 0.06) 60%, transparent 100%);
-				border-color: var(--pfu-color-control-highlight-border);
-				box-shadow: none;
-				color: var(--pfu-color-control-active-content);
-			}
-		}
-
 		.wrapper {
 			padding: 8px;
 			display: flex;
@@ -330,6 +299,10 @@
 			display: flex;
 			flex-direction: row;
 			gap: 8px;
+
+			.search {
+				flex: 2;
+			}
 		}
 
 		.codex-panel {
