+++ templates/actor/party/partials/actor-party-codex-category.hbs
@@ -1,68 +1,3 @@
 <div class="codex-panel {{#unless (eq activeCodexTab category)}}hidden{{/unless}}"
 	 data-panel="{{category}}">
-
-	<div class="toolbar">
-		<input class="search"
-			   type="text"
-			   placeholder="..."
-			   data-codex-search="" />
-		<button class="codex-btn codex-btn--add"
-				data-action="addCodexEntry" data-category="{{category}}">
-			<span>{{localize 'FU.Add'}}</span>
-		</button>
-	</div>
-
-	<ul class="">
-		{{#each (lookup system.codex category) as |entry index|}}
-			<li class="section-container entry" data-index="{{index}}" data-category="{{../category}}">
-
-				<div class="left">
-					<a data-action="forCodexEntry"
-					   data-type="view"
-					   data-index="{{index}}"
-					   title="View"
-					   data-category="{{../category}}">
-						<img class="profile-img flexcol" src="{{lookup entry 'img'}}"
-							 alt="Artwork"/>
-					</a>
-					{{#if @root.isOwner}}
-					<div class="controls">
-
-							<button data-action="forCodexEntry"
-									data-type="edit"
-									data-index="{{index}}"
-									data-category="{{../category}}"
-									title="Edit">✎</button>
-							<button data-action="forCodexEntry"
-									data-type="display"
-									data-index="{{index}}"
-									title="{{localize 'SIDEBAR.CharArt'}}"
-									data-category="{{../category}}">
-								<i class="fu-icon--xs fa-solid fa-image"></i>
-							</button>
-							<button data-action="forCodexEntry"
-									data-type="send"
-									data-index="{{index}}"
-									data-category="{{../category}}">
-								<i class="fu-icon--xs fas fa-comment"></i>
-							</button>
-							<button data-action="removeArrayElement"
-									data-path="system.codex.{{../category}}"
-									data-index="{{index}}"
-									title="FU.Delete">✕</button>
-					</div>
-					{{/if}}
-				</div>
-
-				<div class="right">
-					<span class="name">
-						{{entry.name}}
-					</span>
-					<span class="description">
-						{{{entry.enrichedDescription}}}
-					</span>
-				</div>
-			</li>
-		{{/each}}
-	</ul>
 </div>
